Split train_locally into a binary and a library
We have additional binaries that call into train_eval to take advantage
of internal infrastructure. The recent flag refactoring change causes
issues as if we wanted train_eval we previously had to import the binary
which already defined the flags. This patch makes the structure a lot
cleaner.
